Check MISRA C:2025 and bring the code into conformance
Every lint target also checks MISRA C:2025, and the code base conforms, with 18 deviations recorded in `lint/MISRA-DEVIATIONS.md`, none of them from a mandatory guideline. Predicates and flags become `bool` as part of this, which changes the platform interface.

- `lint/misra.lnt`: PC-lint Plus `au-misra5.lnt`, kept out of library headers, with the deviations; `//lint` comments cite them where they apply to one line or file
- predicates and flags are `bool`: `asdf_put_code`, `asdf_next_code`, `asdf_available`, `asdf_keymap_valid`, the ring, repeat, modifier, physical, and virtual predicates, `asdf_t.repeat_armed`; strobe polarity is `bool` through `asdf_platform_t`
- unsigned constants throughout: matrix limits, keymap flags, masks, and output values; SHIFT, CAPS, and CTRL states are unsigned bit constants rather than enums
- conditions compare explicitly; precedence is parenthesized; no side effects on the right of `&&` or `||`; no chained assignments; parameters are not modified
- `asdf_elapsed_ticks` replaces two copies of the elapsed-time clamp; `asdf_process` sends codes in a counted loop
- `ASDF_KEY` casts each field; `ASDF_IDLE_VALUE` replaces `!ACTIVE_VALUE`; function tables take `&function`
- the keymap generator designates keys by column and compares sizes unsigned; the keymap registry is sized by `create_keymap_slots` and local to `asdf_keymap_descriptor`
- lookup tables move into the one function that uses them; statics no longer share names with struct members or parameters
- PIC32CM: 32-bit masks before shifting, `__NOP()` rather than inline assembly, `_sbrk` advances by indexing
- README and release notes: MISRA checking and the `bool` interface

+++ src/Arch/asdf_arch_pic32cm_pl10_dip28.h
@@ -46,7 +46,7 @@
 // ASCII: 8-bit parallel output on PA08..PA15.
 #define ASCII_GROUP 0u
 #define ASCII_SHIFT 8u
-#define ASCII_MASK (0xFFu << ASCII_SHIFT)
+#define ASCII_MASK ((uint32_t) 0xFFu << ASCII_SHIFT)
 // LEDs (active low) on PA16/PA17/PA18.
 #define LED1_GROUP 0u
 #define LED1_PIN 16u
